fix rest of ts errors and move func to helpers
